+++ BigShow1949/Classes/04 - ViewTransition(视图切换)/FlipPageView/YFFlipPageViewController.m
@@ -65,7 +65,7 @@
         page.backgroundColor = [UIColor redColor];
     }else{}
     
-    page.tempContentLabel.text = [NSString stringWithFormat:@"%d", index];
+    page.tempContentLabel.text = [NSString stringWithFormat:@"%lu", (unsigned long)index];
     
     return page;
 }
